working/requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
crashtest==0.4.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
docopt==0.6.2
dulwich==0.21.7
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
idna==3.10
importlib_metadata==8.5.0
installer==0.7.0
ipython==8.12.3
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==24.3.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodomex==3.21.0
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pyzmq==26.2.0
RapidFuzz==3.10.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.21.0
shellingham==1.5.4
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.4.0
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
trove-classifiers==2024.10.21.16
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
webencodings==0.5.1
xattr==1.1.0
yarg==0.1.9
zipp==3.20.2
